requirements.txt
numpy>=1.25
matplotlib>=3.8
seaborn>=0.12
networkx>=3.0
scipy>=1.10
pillow>=9.0
